backend/requirements.txt
llama-index-llms-openai==0.2.10
llama-index-readers-wikipedia==0.2.0
llama-index-postprocessor-cohere-rerank==0.2.1
llama-index-embeddings-openai==0.2.5
llama-index-agent-openai==0.3.4
nest-asyncio==1.6.0 # latest version
pydantic==2.9.2
llama-parse==0.5.6
cohere==5.11.0
fastapi==0.115.0
uvicorn==0.31.1
pytest==8.3.2
celery==5.4.0
qdrant-client==1.10.1
packaging>=24.1
function_schema>=0.4.5
python-decouple==3.8
redis==5.0.7
matplotlib
seaborn
xgboost
imblearn
firecrawl-py==1.2.4
llama-index-readers-web==0.2.3
# faiss-cpu>=1.8.0.post1
# llama-index==0.11.15 # from 2/10
# llama-index-embeddings-huggingface==0.3.1
# llama-index-vector-stores-faiss==0.2.0
# llama-index-vector-stores-pinecone==0.2.1